requirements.txt
python-binance==1.0.19
pandas==2.0.3
numpy==1.24.3
python-telegram-bot==20.7
python-dotenv==1.0.0
loguru==0.7.2
schedule==1.2.1
requests==2.31.0
discord.py==2.3.2
pandas_ta==0.3.14b0
aiohttp==3.8.5
asyncio==3.4.3
streamlit>=1.29.0
streamlit-autorefresh>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
flask==2.3.3
matplotlib
seaborn
websockets>=12.0
plotly>=5.17.0